+++ compute-budget/src/lib.rs
@@ -2,5 +2,5 @@
 #![cfg_attr(RUSTC_WITH_SPECIALIZATION, feature(min_specialization))]
 
 pub mod compute_budget;
-pub mod compute_budget_processor;
+pub mod compute_budget_limits;
 pub mod prioritization_fee;
